refactor(presenter/interfaces): creating router interfaces folder
