+++ src/app/presenter/models/interfaces/router/route-animation-params.interface.ts
