+++ src/app/presenter/models/interfaces/router/route-data.interface.ts
@@ -1,5 +1,5 @@
 import { UserAccessParams } from "@infra/guards/user-access/user-access-params.interface";
-import { RouteAnimationParams } from "@presenter/models/interfaces/route-animation-params.interface";
+import { RouteAnimationParams } from "@presenter/models/interfaces/router/route-animation-params.interface";
 
 export type RouteData = UserAccessParams & RouteAnimationParams;
